requirements.txt
# Mental Health Predictor - Python Dependencies

# Core ML & Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning (Optional)
torch>=2.0.0
# tensorflow>=2.13.0  # Uncomment if using TensorFlow

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web Frameworks
gradio>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI Enhancement
colorama>=0.4.6  # For Windows color support

# Data Validation
pydantic>=2.0.0

# Logging
loguru>=0.7.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools (Optional)
black>=23.0.0
flake8>=6.0.0
ipython>=8.14.0
jupyter>=1.0.0
